Refactor: move built-in resolvers to built_in_resolvers.py
No functional change
